requirements.txt
watchdog==3.0.0
python-docx==0.8.11
python-pptx==0.6.21
PyPDF2==3.0.1
docx2pdf==0.1.8 